refactored deletePatient method with Logging and a FeatureFlag

+++ PatientService/Controllers/PatientController.cs
@@ -1,13 +1,12 @@
-using System;
-using System.Net.Http;
-using System.Text;
-using System.Text.Json;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
+using Microsoft.IdentityModel.Tokens;
+using Monitoring;
+using OpenTelemetry.Trace;
 using PatientService.Core.DTOs;
 using PatientService.Interfaces;
 using Polly;
+using FeatureHub;
 
 namespace Patient.Controllers;
 
@@ -18,11 +17,15 @@ public class PatientController : ControllerBase // Make sure it derives from Con
     private readonly ILogger<PatientController> _logger;
     private readonly IPatientService _patientService;
     private readonly HttpClient _httpClient = new();
+    private readonly Tracer _tracer;
+    private readonly FeatureHubClient _featureHubClient;
 
-    public PatientController(ILogger<PatientController> logger, IPatientService patientService)
+    public PatientController(ILogger<PatientController> logger, IPatientService patientService, Tracer tracer, FeatureHubClient featureHubClient)
     {
         _logger = logger;
         _patientService = patientService;
+        _tracer = tracer;
+        _featureHubClient = featureHubClient;
     }
     
     [HttpGet]
@@ -74,9 +77,21 @@ public class PatientController : ControllerBase // Make sure it derives from Con
     [Route("deletePatient/{ssn}")]
     public async Task<IActionResult> DeletePatient([FromRoute] int ssn)
     {
+        
+        using var activity = _tracer.StartActiveSpan("DeletePatient");
+        Logging.Log.Information("A deletePatient function has been called.");
+        
         try
         {
+            var country = HttpContext.Request.Headers["country"];
+            if (country.IsNullOrEmpty()) return BadRequest("No country provided");
+            
+            var feature = await _featureHubClient.IsCountryAllowed(country);
+            
+            if (!feature) return StatusCode(403, $"{country} is not part of our service area.");
+
             await _patientService.DeletePatient(ssn);
+            
             return StatusCode(201, "Patient successfully deleted");
         }
         catch (Exception e)
